Refactor syscall dispatcher and enable format string checking (#69)
* Refactoring the syscalls to make it more readable

* Fixed minor bug in if-statement

* Moved truncate

* Moved filesystem syscalls to a seperate file

* Removed ptables_init

* Removed inode_cache_init

* Added print attribute to easily find formatting bugs in printk

+++ kernel/src/fs/ext2/ialloc.c
@@ -165,7 +165,7 @@ int ext2_free_inode(vfs_inode_t* dir)
     unsigned long bit = (dir->i_ino - 1) % es->s_inodes_per_group;
     int oldbit = atomic_clear_bit((s32)bit, (void*)bitmap);
     if (!oldbit) {
-        printk("%s: Race condition on bit %d, retrying\n", __func__, bit);
+        printk("%s: Race condition on bit %d, retrying\n", __func__, (int)bit);
         return -1;
     }
